Deal with rigs in the app file, and slots that arent called vehicle slots.
